tutorials/simple-demo/backend/src/transitions/make_predictions/requirements.txt
lucidtech-las == 3.1.0
backoff
filetype
requests
urllib3
